phase_1/requirements.txt
# Phase 1: Data Collection & Preprocessing

beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
pyyaml==6.0.1
